Add basic structure for simple interpretation
This defines the most basic form of the interpreter pattern: we can
build up any usual console program using these primitives! It does
present a problem, though: we have no way of re-using programs that
we write to build bigger programs. Ideally, we would like to be able to
define parts of programs that we can plug together.