Add the Map instruction and rename EndWith to Pure
The Map instruction can be used to add custom computations and
transformations. Because any given value can be transformed, the EndWith
operation needn't exist in order for us to "end with" the value we want.
However, it can still exist as a way of lifting values into the
computation as "pure" programs, so we rename it to Pure.